+++ Sources/VComponents/Components/Modals/Bottom Sheet/_VBottomSheet.swift
@@ -16,6 +16,7 @@ struct _VBottomSheet<HeaderLabel, Content>: View
 {
     // MARK: Properties
     @Environment(\.presentationHostPresentationMode) private var presentationMode: PresentationHostPresentationMode
+    @StateObject private var interfaceOrientationChangeObserver: InterfaceOrientationChangeObserver = .init()
     
     private let model: VBottomSheetModel
     
@@ -26,7 +27,7 @@ struct _VBottomSheet<HeaderLabel, Content>: View
     private let content: () -> Content
     
     private var hasHeader: Bool { headerLabel.hasLabel || model.misc.dismissType.hasButton }
-    private var hasGrabber: Bool { model.misc.dismissType.contains(.pullDown) || model.layout.height.isResizable }
+    private var hasGrabber: Bool { model.misc.dismissType.contains(.pullDown) || model.layout.sizes.current.size.heights.isResizable }
     private var hasHeaderDivider: Bool { hasHeader && model.layout.headerDividerHeight > 0 }
     
     @State private var isInternallyPresented: Bool = false
@@ -51,7 +52,7 @@ struct _VBottomSheet<HeaderLabel, Content>: View
         self.headerLabel = headerLabel
         self.content = content
         
-        _offset = .init(initialValue: model.layout.height.max - model.layout.height.ideal)
+        _offset = .init(initialValue: model.layout.sizes.current.size.heights.max - model.layout.sizes.current.size.heights.ideal)
     }
 
     // MARK: Body
@@ -68,6 +69,7 @@ struct _VBottomSheet<HeaderLabel, Content>: View
                 of: presentationMode.isExternallyDismissed,
                 perform: { if $0 { animateOutFromExternalDismiss() } }
             )
+            .onChange(of: interfaceOrientationChangeObserver.orientation, perform: { _ in resetHeightFromOrientationChange() })
     }
     
     private var blinding: some View {
@@ -78,13 +80,13 @@ struct _VBottomSheet<HeaderLabel, Content>: View
     }
     
     @ViewBuilder private var bottomSheet: some View {
-        if model.layout.height.isLayoutValid {
+        if model.layout.sizes.current.size.heights.isLayoutValid {
             ZStack(alignment: .top, content: {
                 VSheet(model: model.sheetModel)
                     .if(!model.misc.isContentDraggable, transform: { // NOTE: Frame must come before DragGesture
                         $0
-                            .frame(height: model.layout.height.max)
-                            .offset(y: isInternallyPresented ? offset : model.layout.height.max)
+                            .frame(height: model.layout.sizes.current.size.heights.max)
+                            .offset(y: isInternallyPresented ? offset : model.layout.sizes.current.size.heights.max)
                             .gesture(
                                 DragGesture(minimumDistance: 0)
                                     .onChanged(dragChanged)
@@ -105,14 +107,15 @@ struct _VBottomSheet<HeaderLabel, Content>: View
                     .frame(maxHeight: .infinity, alignment: .top)
                     .if(!model.misc.isContentDraggable, transform: { // NOTE: Frame must come before DragGesture
                         $0
-                            .frame(height: model.layout.height.max)
-                            .offset(y: isInternallyPresented ? offset : model.layout.height.max)
+                            .frame(height: model.layout.sizes.current.size.heights.max)
+                            .offset(y: isInternallyPresented ? offset : model.layout.sizes.current.size.heights.max)
                     })
             })
+                .frame(width: model.layout.sizes.current.size.width)
                 .if(model.misc.isContentDraggable, transform: {  // NOTE: Frame must come before DragGesture
                     $0
-                        .frame(height: model.layout.height.max)
-                        .offset(y: isInternallyPresented ? offset : model.layout.height.max)
+                        .frame(height: model.layout.sizes.current.size.heights.max)
+                        .offset(y: isInternallyPresented ? offset : model.layout.sizes.current.size.heights.max)
                         .gesture(
                             DragGesture(minimumDistance: 0)
                                 .onChanged(dragChanged)
@@ -184,7 +187,7 @@ struct _VBottomSheet<HeaderLabel, Content>: View
     }
 
     private var contentView: some View {
-        ZStack(alignment: .top, content: {
+        ZStack(content: {
             if !model.misc.isContentDraggable {
                 Color.clear
                     .contentShape(Rectangle())
@@ -200,9 +203,10 @@ struct _VBottomSheet<HeaderLabel, Content>: View
                             .frame(height: UIDevice.safeAreaInsetBottom)
                     })
             })
+            .frame(maxWidth: .infinity)
             .if(
                 model.layout.autoresizesContent,
-                ifTransform: { $0.frame(height: model.layout.height.max - offset - grabberHeaderDividerHeight) },
+                ifTransform: { $0.frame(height: model.layout.sizes.current.size.heights.max - offset - grabberHeaderDividerHeight) },
                 elseTransform: { $0.frame(maxHeight: .infinity) }
             )
     }
@@ -276,8 +280,8 @@ struct _VBottomSheet<HeaderLabel, Content>: View
         currentDragValue = dragValue
         
         let newOffset: CGFloat = offsetBeforeDrag + dragValue.translation.height
-        let maxAllowedOffset: CGFloat = model.layout.height.max - model.layout.height.min
-        let minAllowedOffset: CGFloat = model.layout.height.max - model.layout.height.max
+        let maxAllowedOffset: CGFloat = model.layout.sizes.current.size.heights.max - model.layout.sizes.current.size.heights.min
+        let minAllowedOffset: CGFloat = model.layout.sizes.current.size.heights.max - model.layout.sizes.current.size.heights.max
         
         withAnimation(.linear(duration: 0.1), { // Get's rid of stuttering
             offset = {
@@ -314,7 +318,7 @@ struct _VBottomSheet<HeaderLabel, Content>: View
             guard let offsetBeforeDrag = offsetBeforeDrag else { return }
             
             animateOffsetOrPullDismissFromSnapAction(.dragEndedSnapAction(
-                height: model.layout.height,
+                heights: model.layout.sizes.current.size.heights,
                 canPullDownToDismiss: model.misc.dismissType.contains(.pullDown),
                 pullDownDismissDistance: model.layout.pullDownDismissDistance,
                 offset: offset,
@@ -324,7 +328,7 @@ struct _VBottomSheet<HeaderLabel, Content>: View
             
         case true:
             animateOffsetOrPullDismissFromSnapAction(.dragEndedHighVelocitySnapAction(
-                height: model.layout.height,
+                heights: model.layout.sizes.current.size.heights,
                 offset: offset,
                 velocity: dragValue.velocity(inRelationTo: previousDragValue).height
             ))
@@ -336,6 +340,11 @@ struct _VBottomSheet<HeaderLabel, Content>: View
         case .dismiss: animateOutFromDrag()
         case .snap(let newOffset): withAnimation(model.animations.heightSnap, { offset = newOffset })
         }
+    }
+    
+    // MARK: Orientation
+    private func resetHeightFromOrientationChange() {
+        offset = model.layout.sizes.current.size.heights.max - model.layout.sizes.current.size.heights.ideal
     }
 }
 
@@ -365,5 +374,6 @@ struct VBottomSheet_Previews: PreviewProvider {
                     }
                 )
             })
+            .previewInterfaceOrientation(.portrait)
     }
 }