+++ Sources/VComponents/Components/Modals/Modal/VModalModel.swift
@@ -44,11 +44,8 @@ public struct VModalModel {
     /// Sub-model containing layout properties.
     public struct Layout {
         // MARK: Properties
-        /// Modal size. Defaults to `0.9` ratio of screen with and `0.6` ratio of screen height.
-        public var size: CGSize = .init(
-            width: UIScreen.main.bounds.width * 0.9,
-            height: UIScreen.main.bounds.height * 0.6
-        )
+        /// Modal sizes. Defaults to `default`.
+        public var sizes: Sizes = .default
         
         /// Edges ignored by keyboard. Defaults to `[]`.
         public var ignoredKeybordSafeAreaEdges: Edge.Set = []
@@ -83,6 +80,70 @@ public struct VModalModel {
         // MARK: Initializers
         /// Initializes sub-model with default values.
         public init() {}
+        
+        // MARK: Sizes
+        /// Model that describes modal sizes.
+        public struct Sizes {
+            // MARK: Properties
+            /// Portrait size configuration.
+            public let portrait: SizeConfiguration
+            
+            /// Landscape size configuration.
+            public let landscape: SizeConfiguration
+            
+            /// Current size configuration based on interface orientation.
+            public var current: SizeConfiguration {
+                switch _IntefaceOrientation() {
+                case nil: return portrait
+                case .portrait: return portrait
+                case .landscape: return landscape
+                }
+            }
+            
+            // MARK: Initializers
+            /// Initializes `Sizes` with size configurations.
+            public init(portrait: SizeConfiguration, landscape: SizeConfiguration) {
+                self.portrait = portrait
+                self.landscape = landscape
+            }
+            
+            /// Default value. Set to `0.9` ratio of screen width and `0.6` ratio of screen height in portrait, and reverse in landscape.
+            public static var `default`: Sizes {
+                .init(
+                    portrait: .relative(.init(width: 0.9, height: 0.6)),
+                    landscape: .relative(.init(width: 0.6, height: 0.9))
+                )
+            }
+            
+            // MARK: Single Size Configuration
+            /// Enum that describes state, either `point` or `relative`.
+            public enum SizeConfiguration {
+                // MARK: Cases
+                /// Size configuration with fixed sizes represented in points.
+                case point(CGSize)
+                
+                /// Size configuration with ratios relative to screen sizes.
+                case relative(CGSize)
+                
+                // MARK: Properties
+                /// Size configuration represented in points.
+                ///
+                /// `point` configuration is returned directly,
+                /// while `relative` configurations are multiplied by the screen size.
+                public var size: CGSize {
+                    switch self {
+                    case .point(let size):
+                        return size
+                    
+                    case .relative(let size):
+                        return .init(
+                            width: UIScreen.main.bounds.size.width * size.width,
+                            height: UIScreen.main.bounds.size.height * size.height
+                        )
+                    }
+                }
+            }
+        }
         
         // MARK: Margins
         /// Sub-model containing `leading`, `trailing`, `top`, and `bottom` margins.